requirements.txt
streamlit
pandas
plotly
altair
snowflake-connector-python
